backend/requirements.txt
fastapi[all]
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
pymongo
openai
google-generativeai
cohere
anthropic
redis
